lm_sensors: Automatically restart fancontrol
On versal, when the FPGA is erased / programmed / modified, the sensor
can go away and the fancontrol process dies.  We need to auto restart
this so the fan is running properly.

As the system controller software is only supported for systemd use, we
did not modify the sysvinit file.

Signed-off-by: Jonathan Stroud <jonathan.stroud@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit a0b0a5a6b986f7eeb55b407ac94539d9147254d9)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>