Define PACKAGE_UBOOT_DTB_NAME config option
PACKAGE_UBOOT_DTB_NAME is petalinux variable and it's value in petalinux build flow is "uboot-device-tree.dtb"
But this variable is not getting updated in yocto flow while reselected dtb generation,which is causing failure in eval-brd-sc board specific rpm build process.
As a result defined PACKAGE_UBOOT_DTB_NAME variable.

This is needed in case like VEK280-revB,because here we are using ADI-PHY ethernet which is included in uboot-device-tree.dtb.

Signed-off-by: Tejas Bhumkar <tejas.arvind.bhumkar@amd.com>