eval-brd-sc-zynqmp: Remove boot.bin from wic
The boot.bin is included in the QSPI image for this system so it should
not be installed in the wic image. Firmware components should not be
built as a dependency when building images as they are not required.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit b1e15eebf0696a10f3fcfefad6788d2067f147aa)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>