labtool-jtag-support: use systemd-notify to indicate service is ready
Add systemd-notify to let services that depend on this know that the
service is ready.  Also remove the PreExec function that is no longer
needed because the dfx-mgr now also notifies systemd when it is ready.

Signed-off-by: Jonathan Stroud <jonathan.stroud@amd.com>

--
CR: CR-1172157

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 56960aab74c5aa08e514928a13a60ce6115a0753)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>