remove the separate /boot partition
In anticipation of the implementation of the upcoming A/B update strategy
remove the creation of a separate /boot partition. The /boot partition mostly
contains U-Boot artifacts (i.e. the boot script), device-tree files, Linux
kernel(s), etc. In an A/B system these files should be updated atomically with
the root filesystem since there are often dependencies between them. Therefore
combine the contents of /boot with the rest of the filesystem so all build
artifacts are combined together into one partition. There is no technical need
for a separate, vfat-formatted /boot partition.

This change is eval-brd-sc-zynqmp specific.
Tested on the system-controller of the VEK280.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>