systed: add repart PACKAGECONFIG
In anticipation of the implementation of a new, systemd-based repartitioning
and resizing mechanism (based on systemd-repart) add the "repart" component to
systemd.

The "repart" systemd component depends on having "openssl" enabled in systemd,
therefore include that PACKAGECONFIG as well.

This change is eval-brd-sc-zynqmp specific.
Tested on the system-controller of a VEK280 board.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>

PR is update to ensure machine specific version is newer then generic version.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>