repartition and format/resize
Replace the previous, one-off, home-brew resizing mechanism with one based on
systemd-repart.

The eMMC on the VEK280 board, for example, is currently 32GB. The Xilinx Boot
Image Recovery Tool (BIRT) does not allow the user to flash an image that is
larger than 4GB. Even at that, flashing a 2.5GB image takes about 10 minutes,
any sort of image used for an A/B+data setup would be quite large and take a
very long time to flash even if it were possible. With this mechanism were
are able to create and ship small images which then repartition and expand
at boot to setup additional partitions and filesystems based on the provided
configuration.

Unfortunately the systemd-repart mechanism in langdale (i.e. 2023), i.e.
systemd v251, appears to have a bug whereby the unit file runs too early in
the boot, such that the drive's /dev nodes are not setup and the tool has
nothing to work on. Attempts to move the provided unit file elsewhere in the
boot order is met with systemd finding dependency loops in the service file
ordering. Therefore a new service file was created, one that won't conflict
with existing service files.

This has all been tested on a more recent version of systemd (v255) and works
perfectly out-of-the-box with just the configuration files. Therefore at that
point remove the temp-repart service file and place the config files directly
in /etc/repart.d (instead of /etc/t-repart.d). Probably easiest to simply
use a systemd bbappend at that point. The config files for this new service
file were placed in an alternate directory so that we could still install and
use the systemd-repart tool but without triggering it to try running (and
reporting failures during system bring-up).

This change is eval-brd-sc-zynqmp specific.
Tested on the system-controller of a VEK280 board.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 1663a9cf269406b7e3d85e5c93286a280eabdefa)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>