enable rauc
Put the pieces in place to enable rauc A/B updates.

Building for rauc introduces a new target. By default we usually bitbake an
image, but with and A/B update system one only needs an image once: the very
first time a board is flashed. Afterwards an image is no longer used, but
rather the system is updated via an update bundle. In this commit the bundle
is named `system-controller-rauc-bundle` but could be renamed to anything.

This solution disables the petalinux initramfs.

Part of any A/B support logic requires setting and saving some variables into
a non-volatile U-Boot environment storage (i.e. 'env save'). Unfortunately
early versions of this device's U-Boot were configured to store the
environment in the same place where the Xilinx Boot Image Recovery Tool is
stored. If this script blindly does an 'env save' the user's ability to load
a new image through the BIRT might go away if the user is using a new Linux
rootfs image (eMMC) with an older U-Boot (boot.bin in QSPI). Merely checking
the U-Boot version will not suffice; both before and after the fix the U-Boot
version is the same. Therefore in order to detect a new versus old U-Boot the
boot.bin in the QSPI must be examined; if it contains the string 'SC-BootFW'
at offset 0x70 then we know this is a new U-Boot and can proceed with the A/B
logic, otherwise we have to fall back to non A/B behaviour.

Further complicating the situation, there are 2 boot.bins, therefore the code
must start by figuring out which of the two is running. This can be determined
by examining the values in the "Persistent Registers" section (QSPI, mtd2,
0x0020.0000). Data in the "Persistent Registers" section not only describes
which of the two boot.bins is running, but also provides the offsets for
finding both images in the QSPI.

As a further complication there is a "Persistent Registers Backup" section
(QSPI, mtd3, 0x00f8.0000) in case the primary is bad.

Also, U-Boot's shell/scripting language is very primitive. It can't compare
strings directly, it can only compare blocks of RAM to each other. Also there
is no way to place a string into RAM directly (to compare against) so the
compare-to string must be placed in RAM one byte at a time.

*** NOTE: SECURITY ALERT! ***
rauc comes with security built into it: bundles are signed when built, and
verified when installed. However this specific example of how to use rauc
simply re-uses the test keys that are in meta-rauc-community; both the public
and private pieces of which are available publicly. Therefore this example
implementation is not secure! In a "real" deployment, customers would be
expected to generate and use their own keys!
*** NOTE: SECURITY ALERT! ***

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>