apache2: Add update-alternative support
Busybox has the ability to provide an httpd server. By default it is not
configured to do so. In our system-controller builds, busybox is configured to
build/provide/install its httpd server. The software that runs on the system
controller requires a webserver. This code does not work with busybox's httpd
server and requires apache2. The busybox recipe understands that it might not
be the only package providing httpd, therefore its webserver symlink is under
the control of update-alternatives.

Since busybox does not provide an httpd server by default, apache2 does not
expect to have to share control of /usr/sbin/httpd, therefore its recipe
does not contain any configuration to support update-alternatives. As a
result, on the system controller when busybox is updated (dnf upgrade) it
reports an error since the /usr/sbin/httpd object is not a symlink under
update-alternatives' control.

Add update-alternative information to apache2 so it plays nicely with
boxybox which is also providing /usr/sbin/httpd.

NOTE: Since pre-existing installs of apache2 do not already install the
/usr/sbin/httpd file under the controll of update-alternatives, trying to
update apache2 to this newer package (or any future ones) will fail since the
update-alternative system will not know what to do when trying to install a
symlink to a location that is not already a symlink. This will continue to be
true until the /usr/sbin/httpd file is removed manually. Once removed manaully
it will be possible to update apache2 and busybox.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>