libubootenv: Fix NOR flash handling
Bump the version of the code to release v0.3.5 in order to incorporate a fix
to the way NOR flashes are handled. Before this bump the OSPI is written 1
byte at a time, which leads to excessively long write times, and which could
lead to premature burnout of the underlying device.

Commit message from the code:
	Revert "fw_setenv: fix bug when SPI flash write size != sector size"

	This reverts commit 44ecc1c216007272a6f99a104a71c9d410969d9e.

	mtd writesize was errouneously interpreted as maximum allowed size, but
	it is the minimum size. The patch raises performance issues because on
	NOR flashes single bytes are written.

This patch should be removed when updating to 5.0-scarthgap, or beyond.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>