Revert "apache2: Add update-alternative support"
This reverts commit 1f814328efa97916a45f527de7c86f8a400aa034.

It was decided that the issue caused by introducing the update-alternatives
mechanism to apache2 (i.e. having to remove the /usr/sbin/httpd file manually
before any "dnf upgrade") was worse than the earlier issue (i.e. not being
able to upgrade busybox with dnf). Therefore pull out this change at this
point in the release and reconsider potentially for later.

NOTE: the patch that this patch reverts (i.e. to add update-alternatives
      to apache2) was submitted and accepted in upstream
      meta-openembedded/meta-webserver (master branch).

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>