enable full RAUC support for system-controller
Port the RAUC support that was added for the previous system-controller
release (2023.2), to this release (still, only for the system-controller) with
one change: the user can now elect whether or not to add this layer's RAUC
implementation via a boolean parameter:

        SYSTEM_CONTROLLER_RAUC_DEMO = "yes"

If this option is set, the in-layer RAUC example will be used, otherwise it is
not. This gives users the following options without this layer's RAUC example
getting in the way:
    - not use RAUC at all
    - use the RAUC demo implementation from this layer
    - use RAUC, but create their own RAUC configuration

The previous 2024 release, 2024.1, had the re-partitioning + resizing
component in place. However the previous release was based on langdale, and
the langdale version of systemd did not have a fully working version of
systemd-repart. Therefore for the previous release a work-around (temp-repart)
was implemented.

This commit is based on scarthgap, which does have a working sytemd-repart.
Thererfore remove the work-around (temp-repart) and implement repartitioning
and resizing using systemd entirely.

If the user enables this layer's RAUC demo, the root disk will be
repartitioned into an A/B+data setup using systemd-repart. If the user does
not enable the RAUC demo and does not implement a systemd-repart-based
partitioning scheme of their own, then the root disk will not be repartitioned
but the root partition will be expanded to take up any unused disk space. The
same mechanism (systemd-repart) is used in both cases, simply with different
configurations.

+---------------  !SECURITY NOTE! -------------------------------------------+
|   RAUC is designed such that the update bundles have to be signed with a   |
|   key. In order to make this demo work out-of-the-box, a static key is     |
|   generated and placed in this layer so that it can be used to sign during |
|   the build, and used to verify at runtime in the image. Since this is a   |
|   published key, if it is used in production for your builds, anyone could |
|   use this same key to create valid update bundles for your device!        |
|                                                                            |
|   DO NOT USE THIS KEY                                                      |
|                                                                            |
|   Check the RAUC documentation for instructions on generating and using    |
|   your own keys which must not be shared publicly.                         |
+---------------  !SECURITY NOTE! -------------------------------------------+

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>