boot.cmd.raucsc: update BOOT.BIN version check
When the BOOT.BIN version check was created, the latest BOOT.BIN at that time
had a version string that started with:

	<machine>-BootFW...

At this time the latest BOOT.BIN version strings start with:

	<machine>-v<version>...

Which causes the original check to fail.

Instead of potentially forever chasing new string versions, update the string
check to look for older BOOT.BIN version strings and don't perform RAUC if
these old versions are found.

It is my understanding that the BOOT.BIN version strings used the following
formats for the given time periods:
	pre-2023:
		BOOT-...
	2023-2024:
		<machine>-BootFW...
	post-2024:
		<machine>-v<version>...

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>