system-controller-rauc-bundle: make MACHINE-specific
This RAUC implementation was created with only the eval-brd-sc-zynqmp{-sdt} in
mind and tested on only that setup. Configure the bundle recipe such that it
can not be built for any other MACHINE. This will cause less confusion for
users who might successfully build for different MACHINEs and believe RAUC is
viable on other platforms.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>