Add a USB Gadget interface service for the System Controller
New boards have a USB connection between the Versal and the System
Controller. This can be used to emulate a USB thumb drive and a USB
Ethernet controller.  This can be used to load the Versal SD card image
onto the SC and boot the Versal without having to physically image the
SD card.

Signed-off-by: Jonathan Stroud <jonathan.stroud@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>