system-controller-nobootpartition.wks: add part-name
In anticipation of using RAUC, since we are using GPT partitioning, an
easy, straight-forward way to distinguish between partitions is to use
the partition's GPT "part-labels". These can be used on the kernel's
cmdline to indicate the root= target.

Some people define both the A and B in the wks, thus doubling the size
of the *wic file with a redundant image that is very likely to never be
used (it will very likely be overwritten with a B image before it is
ever booted). I prefer to not define the B image here and let systemd's
resize/repart create it on first boot. This is why the wks file only
defines a rootfsA.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Message-Id: <20251027154242.2016488-4-trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>